+++ CHANGELOG.md
@@ -2,6 +2,10 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## [Unreleased]
 
+### Features
+
+* Added `profile` command for measuring program evaluation time
+
 ## v22.4.5
 
 ### Bugfixes